Make our publish and promote roles synchronous in nature
The publish and promote tasks should happen synchronously so Jenkins is able to see
if we have some issue or not.
The commit causes hammer to launch the tasks in a synchronous manner whereas we
exploit the concurrency using Ansible's async task behaviour (Trying to get the best
of both worlds ;) ).

Signed-off-by: Saurabh Badhwar <sbadhwar@redhat.com>