Make sure private networking is up
When running on physical hosts (in contrast to VMs provisioned by
ansible-kvm-host-mgr) we can not be sure networking is OK, so we
have employed linux-system-roles.network role.

But that role might need to install package so it needs to go after
roles that configure yum repos.

And common role is already checking networking sanity, so it needs to go
after that new plain-network role.