Allow the usage of an additional disk to store Satellite data
Provide a way to store most of the Satellite data in a different
disk.

To do so, set `partition_new_storage` to `True` and `storage_pools_disk`
with the value of the disk to be used, and a new VG and the required
LV and partitions will be created.

It uses the `linux-system-roles.storage` Ansible role.

Note: this code can be easily reused to implement the same funcionality
for capsules installations.