feat: experiment: Add a new experiment to test parallel capsule sync
This test will synchronize content from a CDN mirror to the Satellite server,
and then run a parallel content synchronization from the Satellite to the
capsules defined in the inventory.

It will initially test one capsule, then two, then four, and so on until
all the capsules in the inventory are used.

This test will also test how many capsules can be managed by a Satellite
server.