feat(experiment/full_fam.sh): Introduce `experiment/full_fam.sh`
This is a WIP version of what `full.sh` will be when we try to
leverage as much FAM (Foreman Ansible Modules) as possible.

It will evolve over time and more code will be converted.

The main problem right now lies in the measurement analysis, but
I hope that being in the tree will help us.